Small refactor of the n-tier solution
