+++ src/UKHO.SalesCatalogueStub.Api/Startup.cs
@@ -1,3 +1,8 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Reflection;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
@@ -10,17 +15,12 @@ using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
 using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Serialization;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Reflection;
 using UKHO.Logging.EventHubLogProvider;
 using UKHO.SalesCatalogueStub.Api.Configuration;
 using UKHO.SalesCatalogueStub.Api.EF;
-using UKHO.SalesCatalogueStub.Api.EF.Repositories;
 using UKHO.SalesCatalogueStub.Api.Filters;
 using UKHO.SalesCatalogueStub.Api.Middleware;
+using UKHO.SalesCatalogueStub.Api.Services;
 
 namespace UKHO.SalesCatalogueStub.Api
 {
@@ -152,7 +152,7 @@ namespace UKHO.SalesCatalogueStub.Api
             services.AddDbContext<SalesCatalogueStubDbContext>((serviceProvider, options) =>
                 options.UseLazyLoadingProxies().UseSqlServer(dbConnectionString));
 
-            services.AddScoped<IProductEditionRepository, ProductEditionRepository>();
+            services.AddScoped<IProductEditionService, ProductEditionService>();
 
             services.AddHealthChecks();
         }
